+++ components/DocsSidebar/SidebarRegionSelector.tsx
@@ -53,7 +53,7 @@ export default function SidebarRegionSelector({ showInfoTip = true }: SidebarReg
         open={open}
         onOpenChange={setOpen}
       >
-        <SelectTrigger className="!flex !h-8 !w-full !items-center !justify-between !gap-0 !rounded !border !border-[var(--l2-border)] !bg-[var(--l2-background-60)] !px-3 !py-0 !text-sm !text-[var(--l1-foreground)] !shadow-none !outline-none !ring-0 hover:!border-[var(--l1-border)] focus:!ring-0 disabled:!cursor-wait disabled:!opacity-100 [&>span]:!min-w-0 [&>span]:!flex-1 [&>span]:!truncate [&>span]:!text-left [&_svg]:!h-3 [&_svg]:!w-3 [&_svg]:!shrink-0 [&_svg]:!text-[var(--l3-foreground)]">
+        <SelectTrigger className="h-8 w-full rounded border border-[var(--l2-border)] bg-[var(--l2-background-60)] px-3 text-sm text-[var(--l1-foreground)] shadow-none hover:border-[var(--l1-border)] focus:ring-0 disabled:cursor-wait disabled:opacity-100 [&>span]:min-w-0 [&>span]:flex-1 [&>span]:truncate [&>span]:text-left [&_svg]:h-3 [&_svg]:w-3 [&_svg]:text-[var(--l3-foreground)]">
           {isLoading ? (
             <span className="flex items-center gap-2 text-[var(--l3-foreground)]">
               <Loader2 size={12} className="animate-spin" />
@@ -64,20 +64,11 @@ export default function SidebarRegionSelector({ showInfoTip = true }: SidebarReg
           )}
         </SelectTrigger>
         <SelectContent
-          className="!z-[100] !w-[var(--radix-select-trigger-width)] !min-w-[var(--radix-select-trigger-width)] !animate-none !rounded !border !border-[var(--l2-border)] !bg-[var(--l2-background)] !shadow-lg data-[side=bottom]:!animate-none data-[side=top]:!animate-none data-[state=closed]:!animate-none data-[state=open]:!animate-none"
+          className="z-[100] w-[var(--radix-select-trigger-width)] min-w-[var(--radix-select-trigger-width)] rounded border border-[var(--l2-border)] bg-[var(--l2-background)] shadow-lg"
           position="popper"
           side="bottom"
           align="start"
           sideOffset={4}
-          style={
-            {
-              '--select-content-open-animation': 'none',
-              '--select-content-close-animation': 'none',
-              '--select-content-slide-up-animation': 'none',
-              '--select-content-slide-down-animation': 'none',
-              animation: 'none',
-            } as React.CSSProperties
-          }
         >
           {regionOptions.map((option) => {
             const isSelected = option.value === currentValue
@@ -85,7 +76,7 @@ export default function SidebarRegionSelector({ showInfoTip = true }: SidebarReg
               <SelectItem
                 key={option.value}
                 value={option.value}
-                className="!relative !flex !w-full !cursor-pointer !items-center !justify-between !gap-3 !rounded-none !px-3 !py-2 !pr-9 !text-sm !text-[var(--l3-foreground)] !outline-none hover:!bg-[var(--l2-background-hover)] hover:!text-[var(--l1-foreground)] focus:!bg-[var(--l2-background-hover)] focus:!text-[var(--l1-foreground)] data-[state=checked]:!text-[var(--l1-foreground)]"
+                className="relative cursor-pointer px-3 py-2 pr-9 text-sm text-[var(--l3-foreground)] hover:bg-[var(--l2-background-hover)] hover:text-[var(--l1-foreground)] focus:bg-[var(--l2-background-hover)] focus:text-[var(--l1-foreground)] data-[state=checked]:text-[var(--l1-foreground)]"
               >
                 <span className="min-w-0 truncate">{option.label}</span>
                 {isSelected && (
