+++ components/DocsTOC/docLayoutClasses.ts
@@ -2,6 +2,8 @@
 
 /** Pixel width of the docs left sidenav (`w-80`). Keep in sync with the classes below. */
 export const DOC_SIDENAV_WIDTH_PX = 320
+export const DOC_TOC_WIDTH_PX = 320
+export const DOCS_CONTENT_COLUMN_ATTR = 'data-docs-content-column'
 
 export const DOC_TOC_CLASSES =
   'sticky top-[50px] box-border flex h-[calc(100vh-156px)] w-80 min-w-[320px] max-w-[320px] flex-[0_0_320px] flex-col gap-1 overflow-x-hidden overflow-y-hidden px-4 max-lg:!hidden'
