+++ components/NozPeek/NozPeek.css
@@ -27,6 +27,11 @@
   overflow: visible;
 }
 
+/* Mobile-only round CTA; desktop uses .noz-peek-stage. */
+.noz-peek-balloon {
+  display: none;
+}
+
 .noz-peek-input {
   position: relative;
   z-index: 2;
@@ -209,14 +214,64 @@
   }
 }
 
+@media (max-width: 767px) {
+  .noz-peek-dock {
+    width: 56px;
+    max-width: 56px;
+    bottom: 16px;
+    left: auto !important;
+    right: 16px;
+    transform: none;
+  }
+
+  .noz-peek-stage {
+    display: none;
+  }
+
+  .noz-peek-balloon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 56px;
+    height: 56px;
+    border-radius: 9999px;
+    border: none;
+    background: #e5484d;
+    color: #fff;
+    box-shadow: 0 4px 16px rgba(229, 72, 77, 0.35);
+    pointer-events: auto;
+    cursor: pointer;
+    padding: 0;
+    appearance: none;
+    -webkit-appearance: none;
+  }
+
+  .noz-peek-balloon:focus-visible {
+    outline: 2px solid var(--accent-primary);
+    outline-offset: 2px;
+  }
+
+  .noz-peek-balloon.animate {
+    animation: noz-input-rise 0.6s cubic-bezier(0.16, 1, 0.3, 1) forwards;
+  }
+
+  .noz-peek-balloon.settled {
+    transform: none;
+    filter: none;
+    opacity: 1;
+  }
+}
+
 @media (prefers-reduced-motion: reduce) {
   .noz-peek-input.animate,
+  .noz-peek-balloon.animate,
   .noz-peek-body.animate,
   .noz-peek-body.animate .noz-peek-head,
   .noz-peek-dock.animate .noz-peek-hands {
     animation: none !important;
   }
-  .noz-peek-input.animate {
+  .noz-peek-input.animate,
+  .noz-peek-balloon.animate {
     transform: none;
     filter: none;
     opacity: 1;
