+++ components/icons/GlobeCheck.tsx
@@ -1,0 +1,28 @@
+import type { SVGProps } from 'react'
+
+/** Lucide `globe-check` icon (not yet in lucide-react@0.540). */
+export default function GlobeCheck({
+  size = 24,
+  className,
+  ...props
+}: SVGProps<SVGSVGElement> & { size?: number }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={size}
+      height={size}
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className={className}
+      aria-hidden
+      {...props}
+    >
+      <path d="m15 6 2 2 4-4" />
+      <path d="M2 12h20A10 10 0 1 1 12 2a14.5 14.5 0 0 0 0 20 14.5 14.5 0 0 0 4-10" />
+    </svg>
+  )
+}
